backend/requirements_minimal.txt
# Essential FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# API and HTTP clients
aiohttp==3.9.1
requests==2.31.0

# Data processing
pandas==2.1.3
numpy==1.24.4

# Web scraping
beautifulsoup4==4.12.2
feedparser==6.0.10

# Text processing (basic)
textblob==0.17.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI Services
google-generativeai==0.3.0